Graphs. Booyeah.
Now there is a second pane in the dropdown for each metric on the dashboard. The left pane is the analytics as normal, while the right pane is a graph of the metric's value.

Initially, number and time metrics will be line charts, while word metrics
will have bar charts of the incidences of the most common words.

Still lots to do in visualization, but this is a good start!

(piggybacked a fix for a NoneTypeError in multi_analytics here too)