More memcache!
1) Memcache for all datapoint queries. We were finally pushing our appengine database quota to the max, so this is necessary now. Now we don't go to the database for datapoints unless we absolutely have to.

This meant jumping through several hoops, which I think you'll see if you look at the "lt_key" and lt_query_key() functions.

2) Caching "most recent datapoint" for each query. This was our bottleneck on /dashboard. Request times go from 6 seconds before to 300 ms with this caching.

It should be noted here that we're now caching at multiple levels. 1) Caching database queries. 2) Caching analytics results. Both are dirtied whenever that metric is updated. They may interfere with eachother in the future, so we should watch this.

Also removed a bunch of expiry times when setting keys. That's not necessary.

Also fixed a bug on the manage data page. We were checking the wrong key for
metric_last_update.