Invite Codes!
Now we can create an invite code (in the database), which we can send to x people, and it will whitelist them for the system.

As part of this, we no longer have a whitelist file. Instead, there's another column in the user object that says is_whitelisted.

The logic in lthandler and home.py is still a bit confusing. Refactor later.